Extract Signal Dict to data_struct.py
