+++ magia/data_struct.py
@@ -1,0 +1,34 @@
+from __future__ import annotations
+
+from collections import UserDict
+
+
+class SignalDict(UserDict):
+    """
+    Signal Dict contains a dictionary of signals keyed by their name / specific alias.
+
+    They are read only after being assigned.
+    Alias (i.e. key of the dict) cannot start with `_`.
+    """
+
+    def __getattr__(self, alias):
+        if alias.startswith("_"):
+            return super().__getattribute__(alias)
+        if alias not in self.data:
+            raise KeyError(f"Signal {alias} is not defined.")
+        return self.data[alias]
+
+    def __getitem__(self, alias):
+        if alias not in self.data:
+            raise KeyError(f"Signal {alias} is not defined.")
+        return self.data[alias]
+
+    def __setitem__(self, alias, value):
+        cur = self.data.get(alias)
+        if value is not None and not hasattr(value, "signal_config"):
+            raise KeyError(f"Object {alias} is not a Signal.")
+        if cur is not None and value is not cur:
+            raise KeyError(f"Signal {alias} is read only. Are you trying to connect it with <<= Operator?")
+
+        if value is not None:
+            self.data[alias] = value
