+++ magia/memory.py
@@ -6,7 +6,8 @@ from dataclasses import dataclass
 from itertools import count
 
 from .constants import SignalType
-from .core import Signal, SignalDict, Synthesizable
+from .core import Signal, Synthesizable
+from .data_struct import SignalDict
 from .io_signal import Input, Output
